+++ examples/src/raise.mli
@@ -1,0 +1,3 @@
+val unwrap : 'a option -> 'a [@@may_raise] [@@pure]
+
+val print : string -> unit [@@may_raise]